[iOS] Introduce BluetoothChooserIOS and BluetoothDeviceList
This CL introduces BluetoothChooserIOS and BluetoothDeviceList
and supports a basic functionality for choosing a bluetooth device.

BluetoothChooserIOS has BluetoothDeviceList that holds UITableView
to manage a device list. UITableView is loaded as a popover and a
device is added to UITableView or updated when
BluetoothChooserIOS::AddOrUpdateDevice() is called.

Bug: 1431447
Change-Id: Ie243c97dae0f40549772f9918e12bd6fabb23e25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4430271
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143764}