[Office upload] Update errors when Drive is not mounted
We can detect when Drive is not mounted, which can be because Drive
is explicitly disabled in Settings. Prioritise this error over the
offline error for the fallback dialog, because offline checks
whether Drive is connected, which can lead to the offline error
showing when the device is actually online.

Clean up some other categories of fallback errors that are no
longer used.

Bug: b/281773568
Change-Id: I6edf42403aee7bdf8f811dc39fc77c26556b197b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4520494
Reviewed-by: Cassy Chun-Crogan <cassycc@google.com>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143833}