arc: Replace ARC vmm swap feature flag
This CL remove the ArcVmmSwap feature flag, and replace it by
policy and shortcut feature flag. Background: Previously the flag
is a platform type flag and read by concierge. Now the vmm_swap
already be added as VM launch parameter, so the PLATFORM_FEATURE
is not necessary.

TEST=ArcVmClientAdapterTest

Bug: b:281767771
Change-Id: Ia8e9de4fd78a33921c6b92bed3cddf9e0ddc8d2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4525222
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Shengsong Tan <sstan@google.com>
Cr-Commit-Position: refs/heads/main@{#1143848}