[Files]: Renames NESTED_SHARED_WITH_ME to INDIRECTLY_SHARED_WITH_ME.
Unlike SHARE and SHARED_WITH_ME files, NESTED_SHARED_WITH_ME do not have
shared metadata bit set on them. This creates a confusing code that,
when read by developers, seems to have shared files that are not
considered shared by the fake DriveFs. In this CL we make it explicit
that these files are not marked as shared.

Change-Id: I2c0451aa2ce056dab3ec6279357dbb9a771ab989
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4531165
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Commit-Queue: Bo Majewski <majewski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143954}