[NaCL] Eliminate support for PPB_VideoDecoder(Dev) API
In 112, we turned off support for the PPB_VideoDecoder(Dev) API with a
Finch killswitch and enterprise escape hatch policy to re-enable it if
needed through 114. 114 has now branched and there have been no
invocations of the API (per the metric that we added), as well as no
fires caused by turning it off.

This CL eliminates support for the API entirely. In particular, we
do the following:

- Eliminate the Finch killswitch and all code guarded by it
- Eliminate support for the enterprise policy
- Remove all associated tests and metrics

The primary benefit is the elimination of ppb_video_decoder_impl.*,
which is one of the remaining users of the legacy VDA IPC interface.
This elimination (paired with elimination of the other remaining user)
will in turn unblock refactorings to eliminate GLImage and legacy
mailboxes from the codebase.

Note: The API itself is still exposed as of now, but no longer
implemented - resource creation will always fail. We leave removal of
the API entirely for a followup project.

OBSOLETE_HISTOGRAM[NaCl.ResourceCreationImpl.CreateVideoDecoderDev_Invoked]=No longer used

Change-Id: I313168875ddb41d56d8a31cab1c5287c0c4b87bd
Bug: 1408421, 1382469
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4135117
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Derek Schuff <dschuff@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143956}