Avoid unnecessary re-creation of ContainerQueryEvaluator
Changing to/from display:contents we used to require a re-creation of
the ContainerQueryEvaluator, but now we should update queries and
invalidate properly in UpdateStyleForNonEligibleContainer().

This also fixes a bug with not invalidating querying elements at all if
we were only using style() queries and display changed to 'contents'.

Bug: 1445044
Change-Id: I1c70ed0ef4e6449e44651021dea49240a80f32e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4526080
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143960}