[v8][pku] Add histogram for CFI page mismatch
We assume a maximum OS page size at compile time and need to bail out
of memory protections if there's a mismatch to the size observed at
runtime.
This histogram will be used to see if this is happening in practice or
if we can turn it into a CHECK instead.

Change-Id: I6c76f4c8e45a32be610d5a91fe92699e75fcc54b
Bug: v8:13355
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4530647
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#1143971}