[UPMCleanup] Only repeat removals on the built-in backend
Deletions are executed on both backends to ensure that data removal
from the built-in backend is immediate and doesn't depend on sync.

For the most part, the built-in backend is the shadow backend.
However, in the case of unenrollments, despite sync being enabled,
the built-in backend becomes the main backend and the android backend
becomes the shadow backend.

This CL ensures that we only repeat deletions on the shadow backend
if that is the built in backend.

Bug: 1439191
Change-Id: If1ce32778522741bacfe2343e8cfd0f44b368e07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4517409
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1143974}