[UPMLocalPwd] Complete the first page of the  warning UI
The text about the migration, including the Chrome channel, is added
as the subtitle. There is a button that will be used to acknowledge the
notice and another to provide more options. They are not active yet.

Bug: 1439761
Change-Id: Iabe06c996f07ae743e56113c0b22821f008f62b4
Low-Coverage-Reason: not useful (The coordinator is not tested because it just initializes the mediator and sets up model change processors.)
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4517415
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1144000}