[OOBE][ChromeVox] Fix issue in test
WelcomeScreenChromeVoxHintTest.VoicesChanged should not wait for the
whole OOBE to load before setting up the ChromeVox extension to simulate
a failure. Instead, it should prepare the extension as soon as possible
and then wait for the Welcome screen in order to run its checks.

Bug: b:282665168, b:266939445
Change-Id: Ie044f2d5bb62eeeb2715f01a4ca8afd7d1c4db34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4529623
Reviewed-by: Osama Fathy <osamafathy@google.com>
Commit-Queue: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/main@{#1144025}