[iOS] Fix crash when opening link in password issues from Safety Check
A crash was occurring when an URL was opened from a password issues
page, when accessed through Safety Check -> Password Checkup. This was
due to the fact that the SafetyCheckCoordinator's
PasswordCheckupCoordinator wasn't cleaned up in the `stop` method of the
SafetyCheckCoordinator. This CL fixes the issue by cleaning up the
PasswordCheckupCoordinator property.

Fixed: 1444543
Change-Id: Ib305adee83ab1c81e2a0ae4a5094ee717b8067a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4529276
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Auto-Submit: Noémie St-Onge <noemies@google.com>
Cr-Commit-Position: refs/heads/main@{#1144035}