Remove rationalization from FieldTypeFromAutocompleteAttributeValue
Before this CL, FieldTypeFromAutocompleteAttributeValue took the string
value of the autocomplete and the max-length attribute of a form control
to return a HtmlFieldType. It contained heuristics to use the max-length
attribute to generate rationalized HtmlFieldTypes. For example, an
<input autocomplete="cc-exp-year" max-length="4"> was mapped to an
HtmlFieldType::kCreditCardExp4DigitYear.

The implementation came with two drawbacks:
1) The max-length had to be carried through many function calls.
2) The heuristic cannot be smarter. E.g. an <input
   autocomplete="cc-exp-year" placeholder="YYYY"> could not be mapped to
   HtmlFieldType::kCreditCardExp4DigitYear.

By splitting the rationalization from the parsing, we can improve this.

The new rationalization is implemented in https://chromium-review.googlesource.com/c/chromium/src/+/4493243 but more will follow.

Bug: 1441057
Change-Id: I04ac191fb73be858e7d6d4af5ab2430eb9d467ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4489455
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Dominic Battre <battre@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1144054}