[Android][CredMan][Mfill] Wire conditional Ui reentry button to delegate
With this CL, the conditional UI reentry point actually shows when an
authentication flow is pending and a password field is focused.
Tapping it allows to complete the authentication flow.

At the moment, it's always shown if a CredMan flow is pending its
completion and only hidden by the flag.
Follow-up CLs will restrict that to pwd/webauthn fields.

Bug: 1444418
Change-Id: I0e1a67d0a4e4c8e07fa548a844a2e2c5990462b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4517535
Commit-Queue: Friedrich Horschig <fhorschig@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144069}