FSA: Exclusively lock swap files
Currently, creating a writable file stream takes a shared lock on the
file, with the idea that multiple writable streams may be associated
with the same file. However, writable streams are expected to be
siloed; creating multiple writers should create a swap file per writer

However, until this CL, the swap file itself was completely unlocked.
Since the swap file lives alongside the file (in the same directory),
sites with access to the parent directory could acquire a file handle
to the swap file directly. This leads to unexpected behavior, such as
the ability to create a writable stream from a swap file.

Ideally this would be tested in WPTs, but swap files are
implementation-defined and it's unclear whether it should even be
possible to acquire a handle to one in the first place.

Bug: 1382215, 1413443, 1439179
Change-Id: Ic8e0897850e437160c1076371bf6f00e3e0fe7c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4521555
Reviewed-by: Daseul Lee <dslee@chromium.org>
Auto-Submit: Austin Sullivan <asully@chromium.org>
Commit-Queue: Daseul Lee <dslee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144079}