Reland "[wayland] Cancel dnd when drag target or dragged window destroyed"
This is a reland of commit 0d9b005cd03fcea1466c8ee5e7f600690a8e380e

See diff against patchset 1 for changes to the original CL.

The original CL was reverted due to flaky failures of the following
test case
  WaylandWindowDragControllerTest.
  HandleDraggedWindowDestructionAfterMoveLoop

When `move_loop_handler->RunMoveLoop({})` is called it is expected
the server generates a wl_data_device.data_offer and
wl_data_device.enter event. The `drag_target_window_` is set during
a drag when the enter event is received.

The issue was that sometimes ``move_loop_handler->RunMoveLoop({})``
would return before the client had received the offer and enter
events, and sometimes not.

If there is no enter received by the client, the target window is
not set - and deleting the target window will not cancel the drag.
When the events arrive later in this case the offer / enter events
are correctly no-oped as the window does not exist.

The inconsistency of delivered events for the test made it impossible
to test without flakiness. This CL added a call to `wl::SyncDisplay()` and updated expectations for what occurs when the target window is
destroyed.

Original change's description:
> [wayland] Cancel dnd when drag target or dragged window destroyed
>
> This CL updates the WaylandWindowDragController to cancel the dnd
> session client-side when either of the drag target window or the
> dragged window are destroyed.
>
> The dnd session is not cancelled when either the pointer grab or
> origin window are destroyed as this occurrence is already supported
> and defined by the dnd code.
>
> The client terminates the dnd session by calling OnDataSourceFinish()
> which resets the wl_data_source and performs some client-side
> cleanup. The spec indicates that the destruction of the data source
> can be performed to cancel the dnd session.
>
> To handle any possible events sent by the server, but not yet
> received by the client after the dnd session has been cancelled, the
> client will no-op any further dnd events after this has occurred.
>
> Bug: b/278482149
> Change-Id: I9f4af789b5392d2c589b31ec08077079c2b18daa
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4502756
> Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
> Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1142496}

Bug: b/278482149
Change-Id: Iec19e02aa34bb93cf291e761fc8f78a67fd52c69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518934
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144095}