[iOS] Fix SigninPromoView UIButtonConfiguration Usage
SigninPromoView primaryButton was refactored to use a
UIButtonConfiguration, but was not correctly implemented before. These
changes fix the issues behind a feature flag. Some notable parts of
this refactor:

- titleLabel properties such as adjustsFontSizeToFitWidth and
minimumScaleFactor are automatically taken into account when using
`attributedTitle` in UIButtonConfiguration which is why they are
omitted in the refactored version of code behind the feature flag

- Parts of configuring UIButtons have been moved from the view to
the configurator because initializing a font needs to be done when
setting the title. Setting the font with a nil NSString will result
in a fatal crash

- UIButton.configuration is initialized in the view and then
manipulated in the configurator

Bug: 1432213
Change-Id: I2e2adf0e9471511279b3d1e99f9e9dcb114432ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4519244
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144101}