Allow price tracking utils to be called from shopping service
This patch allows some of the price tracking utils to be called from
the shopping service. This dramatically simplifies their usage as
clients only need to have access to the shopping service and not the
numerous other parameters (all of which are contained by the service
itself).

Change-Id: I3061576cbdf75937ce29866788942a7f387465bb
Low-Coverage-Reason: The static version is well-tested (these are passthrough).
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4509548
Reviewed-by: Zhiyuan Cai <zhiyuancai@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144112}