Hue component should be powerless for fully transparent colors
A new day, a new time to delete
color-computed-color-mix-function-expected.txt

The spec could have been read somewhat ambiguously in the past. Hue
angles are not premultiplied and therefore are affected more when
mixed with very nearly transparent colors. This produces the somewhat
unexpected effect that mixing with a fully transparent rgb color
will modify the hue angle, but nothing else in the color.

Discussion here clarified that when converting a fully transparent
rectangular color to a polar color space, the resulting hue angle
should be "none":
https://github.com/w3c/csswg-drafts/issues/8612

Bug: 1445400
Change-Id: I62e239578a16f3e625033f9e8874bd6259ca74b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4530766
Reviewed-by: Yi Xu <yiyix@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144117}