[iOS] Inject the frame id for each ContentWebFrame
In ios/web/content, frame ids are chosen on the browser side. The
approach of choosing ids on the JavaScript side, as happens in
WKWebView-based ios/web, cannot be used since messages do not arrive
with any identifier for the RenderFrame that sent them. Instead,
the browser chooses the id for each RenderFrameHost and then needs
to ensure that JavaScript uses that id.

This CL handles this by injecting a replacement for the existing
__gCrWeb.message.getFrameId() function that returns the chosen id.

This CL also moves the logic for telling observers that a WebFrame
became available from DidFinishNavigation to DOMContentLoaded, since
DidFinishNavigation is too early -- this happens when the navigation
is committed, which is early enough that JavaScript executed at this
point will race with document start scripts, and sometimes run before
these scripts (at which point __gCrWeb has not yet been set up).

Bug: 1423527
Change-Id: I0bb5f293330778e06e2f2ac3d796e03fe571275e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4520982
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144120}