[PWA] Create skeleton class for collecting UKMs for ML promotion
This CL creates the skeleton class for collecting UKMs that will be used in triggering the Segmentation ML model for installationpromotion. The design is subject to change in the future, but in the interim it serves as a place where all UKMs will be measured and queried. It is also connected to the AppBannerManager and observes the web_contents for favicon changes and other requirements.

Bug: b/280454826
Change-Id: I42ba2911af178892631bd2823cf4f8a134a1d220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4525177
Reviewed-by: Ella Ge <eirage@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Auto-Submit: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144145}