Deflake ParallelDownloadTest.MiddleSliceDelayedError
The test often fails with timeout.
The test attempts to achieve following order of events:
1. File is being downloaded in three slices, meaining three parallel
requests.
2. Third request completes successfully first.
3. Second request receives headers (not a strong requirement, may
predate 2.).
4. First request completes its own slice and slice that belongs to the
second request.
5. Second request receives full response, but fails to write the data.
6. [Verification] Download is completed successfully, since first
request didn't fail to write and there is a special logic that allows
preceding request complete the work of their successors (but not
otherwise).

The problem is how test waits for 4. and how write failure for 5. is
implemented: second write at offset that belongs to second request
fails. Wait for 4. wasn't done in a correct way: it simply waits for
response being sent from the server, while test needs to wait until it
is processed by the client including file write. So, sometimes first
request fails instead of second and that causes timeout.

This CL makes waiting condition more accurate: until whole file is
written. To ensure that test still makes sense, there is a check that
download is not marked as completed until second request is processed.

Bug: 1432822, 1444105
Change-Id: I1129d80f5268c7992725dff7d15ea2a5223c5edb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4533159
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Cr-Commit-Position: refs/heads/main@{#1144151}