Multiplanar: Add UMA for GmbOutputFormat in media
This change adds an UMA metric in GpuVideoAcceleratorFactoriesImpl
to track which OutputFormats are used on which platforms when creating
VideoFrames and accordingly which of these formats create GMBs more
than 1. This is useful for multiplanar shared images where we will have
single SharedImage for multiplanar formats and accordingly a single
shared image for multiplanar GMBs. So if there are platforms where
we have more than 1 GMBs then we need to have a plan of handling
them together.

Bug: 1366495
Change-Id: Ia10e46d56951498120fd2efcdc6b526c7688043c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4519011
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144169}