[iOS] Remove BrowserLauncher
BrowserLauncher's sole remaining purpose is to provide a BrowserState to
AppState for saving cookies when backgrounding. AppState already has
a BrowserState (via its public MainBrowserState property), so this seems
clearly redundant.

This is actually a functional change, however, although it appears to be
a bug fix. AppState used the "current browser"'s browser state to get
the URLRequestContext to clear cookies. The "current browser" might be
either regular or incognito, so presumably in some circumstances this
is saving the wrong cookies.

In addition, AppState was not actually using the MainAppDelegate param
it was initialized with. This reduces the dependencies for AppState to
just StartupInformation.

Beyond this, a lot of the cleanup in this CL is adding an explicit
include for browser_provider_interface.h for the many many files that
relied on getting it transitively from browser_launcher.h.

All of the tests that create or fake AppState are updated to use the
new init, removing many instances of mocked BrowserLaunchers and
MainApplicationDelegates that were, or are now, totally spurious.

(Some unit tests had C++ style updates as well).

Change-Id: I86cff066cc8ddc3a3c833f092dfdee328657e24d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4502617
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144188}